partial assert tests
