+++ test/UnitTests/MSTest.Core.Unit.Tests/Assertions/AssertTests.cs
@@ -18,7 +18,7 @@ using TestFrameworkV2 = FrameworkV2.Microsoft.VisualStudio.TestTools.UnitTesting
 using TestMethod = FrameworkV1::Microsoft.VisualStudio.TestTools.UnitTesting.TestMethodAttribute;
 
 [TestClass]
-public class AssertTests
+public partial class AssertTests
 {
     #region That tests
 
@@ -32,200 +32,6 @@ public class AssertTests
     public void ThatShouldCacheAssertInstance()
     {
         Assert.AreEqual(TestFrameworkV2.Assert.That, TestFrameworkV2.Assert.That);
-    }
-
-    #endregion
-
-    #region ThrowsException tests
-
-    [TestMethod]
-    public void ThrowsExceptionWithLambdaExpressionsShouldThrowAssertionOnNoException()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.ThrowsException<ArgumentException>(() => { }));
-
-        Assert.IsNotNull(ex);
-        Assert.AreEqual(typeof(TestFrameworkV2.AssertFailedException), ex.GetType());
-        StringAssert.Contains(ex.Message, "Assert.ThrowsException failed. No exception thrown. ArgumentException exception was expected.");
-    }
-
-    [TestMethod]
-    public void ThrowsExceptionWithLambdaExpressionsShouldThrowAssertionOnWrongException()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.ThrowsException<ArgumentException>(
-             () =>
-             {
-                 throw new FormatException();
-             }));
-
-        Assert.IsNotNull(ex);
-        Assert.AreEqual(typeof(TestFrameworkV2.AssertFailedException), ex.GetType());
-        StringAssert.Contains(ex.Message, "Assert.ThrowsException failed. Threw exception FormatException, but exception ArgumentException was expected.");
-    }
-
-    #endregion
-
-    #region ThrowsExceptionAsync tests.
-
-    [TestMethod]
-    public async Task ThrowsExceptionAsyncShouldNotThrowAssertionOnRightException()
-    {
-        Task t = TestFrameworkV2.Assert.ThrowsExceptionAsync<ArgumentException>(
-            async () =>
-                {
-                    await Task.Delay(5).ConfigureAwait(false);
-                    throw new ArgumentException();
-                });
-
-        // Should not throw an exception.
-        await t.ConfigureAwait(false);
-    }
-
-    [TestMethod]
-    public void ThrowsExceptionAsyncShouldThrowAssertionOnNoException()
-    {
-        Task t = TestFrameworkV2.Assert.ThrowsExceptionAsync<ArgumentException>(
-            async () =>
-            {
-                await Task.Delay(5).ConfigureAwait(false);
-            });
-        var ex = ActionUtility.PerformActionAndReturnException(() => t.Wait());
-
-        Assert.IsNotNull(ex);
-
-        var innerException = ex.InnerException;
-
-        Assert.IsNotNull(innerException);
-        Assert.AreEqual(typeof(TestFrameworkV2.AssertFailedException), innerException.GetType());
-        StringAssert.Contains(innerException.Message, "Assert.ThrowsException failed. No exception thrown. ArgumentException exception was expected.");
-    }
-
-    [TestMethod]
-    public void ThrowsExceptionAsyncShouldThrowAssertionOnWrongException()
-    {
-        Task t = TestFrameworkV2.Assert.ThrowsExceptionAsync<ArgumentException>(
-            async () =>
-            {
-                await Task.Delay(5).ConfigureAwait(false);
-                throw new FormatException();
-            });
-        var ex = ActionUtility.PerformActionAndReturnException(() => t.Wait());
-
-        Assert.IsNotNull(ex);
-
-        var innerException = ex.InnerException;
-
-        Assert.IsNotNull(innerException);
-        Assert.AreEqual(typeof(TestFrameworkV2.AssertFailedException), innerException.GetType());
-        StringAssert.Contains(innerException.Message, "Assert.ThrowsException failed. Threw exception FormatException, but exception ArgumentException was expected.");
-    }
-
-    [TestMethod]
-    public void ThrowsExceptionAsyncWithMessageShouldThrowAssertionOnNoException()
-    {
-        Task t = TestFrameworkV2.Assert.ThrowsExceptionAsync<ArgumentException>(
-            async () =>
-                {
-                    await Task.Delay(5).ConfigureAwait(false);
-                },
-            "The world is not on fire.");
-        var ex = ActionUtility.PerformActionAndReturnException(() => t.Wait());
-
-        Assert.IsNotNull(ex);
-
-        var innerException = ex.InnerException;
-
-        Assert.IsNotNull(innerException);
-        Assert.AreEqual(typeof(TestFrameworkV2.AssertFailedException), innerException.GetType());
-        StringAssert.Contains(innerException.Message, "Assert.ThrowsException failed. No exception thrown. ArgumentException exception was expected. The world is not on fire.");
-    }
-
-    [TestMethod]
-    public void ThrowsExceptionAsyncWithMessageShouldThrowAssertionOnWrongException()
-    {
-        Task t = TestFrameworkV2.Assert.ThrowsExceptionAsync<ArgumentException>(
-            async () =>
-            {
-                await Task.Delay(5).ConfigureAwait(false);
-                throw new FormatException();
-            },
-            "Happily ever after.");
-        var ex = ActionUtility.PerformActionAndReturnException(() => t.Wait());
-
-        Assert.IsNotNull(ex);
-
-        var innerException = ex.InnerException;
-
-        Assert.IsNotNull(innerException);
-        Assert.AreEqual(typeof(TestFrameworkV2.AssertFailedException), innerException.GetType());
-        StringAssert.Contains(innerException.Message, "Assert.ThrowsException failed. Threw exception FormatException, but exception ArgumentException was expected. Happily ever after.");
-    }
-
-    [TestMethod]
-    public void ThrowsExceptionAsyncWithMessageAndParamsShouldThrowOnNullAction()
-    {
-        static void a()
-        {
-            Task t = TestFrameworkV2.Assert.ThrowsExceptionAsync<ArgumentException>(null, null, null);
-            t.Wait();
-        }
-        ActionUtility.ActionShouldThrowInnerExceptionOfType(a, typeof(ArgumentNullException));
-    }
-
-    [TestMethod]
-    public void ThrowsExceptionAsyncWithMessageAndParamsShouldThrowOnNullMessage()
-    {
-        static void a()
-        {
-            Task t = TestFrameworkV2.Assert.ThrowsExceptionAsync<ArgumentException>(async () => { await Task.FromResult(true).ConfigureAwait(false); }, null, null);
-            t.Wait();
-        }
-        ActionUtility.ActionShouldThrowInnerExceptionOfType(a, typeof(ArgumentNullException));
-    }
-
-    [TestMethod]
-    public void ThrowsExceptionAsyncWithMessageAndParamsShouldThrowAssertionOnNoException()
-    {
-        Task t = TestFrameworkV2.Assert.ThrowsExceptionAsync<ArgumentException>(
-            async () =>
-            {
-                await Task.Delay(5).ConfigureAwait(false);
-            },
-            "The world is not on fire {0}.{1}-{2}.",
-            "ta",
-            "da",
-            123);
-        var ex = ActionUtility.PerformActionAndReturnException(() => t.Wait());
-
-        Assert.IsNotNull(ex);
-
-        var innerException = ex.InnerException;
-
-        Assert.IsNotNull(innerException);
-        Assert.AreEqual(typeof(TestFrameworkV2.AssertFailedException), innerException.GetType());
-        StringAssert.Contains(innerException.Message, "Assert.ThrowsException failed. No exception thrown. ArgumentException exception was expected. The world is not on fire ta.da-123.");
-    }
-
-    [TestMethod]
-    public void ThrowsExceptionAsyncWithMessageAndParamsShouldThrowAssertionOnWrongException()
-    {
-        Task t = TestFrameworkV2.Assert.ThrowsExceptionAsync<ArgumentException>(
-            async () =>
-            {
-                await Task.Delay(5).ConfigureAwait(false);
-                throw new FormatException();
-            },
-            "Happily ever after. {0} {1}.",
-            "The",
-            "End");
-        var ex = ActionUtility.PerformActionAndReturnException(() => t.Wait());
-
-        Assert.IsNotNull(ex);
-
-        var innerException = ex.InnerException;
-
-        Assert.IsNotNull(innerException);
-        Assert.AreEqual(typeof(TestFrameworkV2.AssertFailedException), innerException.GetType());
-        StringAssert.Contains(innerException.Message, "Assert.ThrowsException failed. Threw exception FormatException, but exception ArgumentException was expected. Happily ever after. The End.");
     }
 
     #endregion
@@ -247,422 +53,6 @@ public class AssertTests
 
     #endregion
 
-    #region InstanceOfType tests
-
-    [TestMethod]
-    public void InstanceOfTypeShouldFailWhenValueIsNull()
-    {
-        static void action() => TestFrameworkV2.Assert.IsInstanceOfType(null, typeof(AssertTests));
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void InstanceOfTypeShouldFailWhenTypeIsNull()
-    {
-        static void action() => TestFrameworkV2.Assert.IsInstanceOfType(5, null);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void InstanceOfTypeShouldPassOnSameInstance()
-    {
-        TestFrameworkV2.Assert.IsInstanceOfType(5, typeof(int));
-    }
-
-    [TestMethod]
-    public void InstanceOfTypeShouldPassOnHigherInstance()
-    {
-        TestFrameworkV2.Assert.IsInstanceOfType(5, typeof(object));
-    }
-
-    [TestMethod]
-    public void InstanceNotOfTypeShouldFailWhenValueIsNull()
-    {
-        Action action = () => TestFrameworkV2.Assert.IsNotInstanceOfType(null, typeof(AssertTests));
-    }
-
-    [TestMethod]
-    public void InstanceNotOfTypeShouldFailWhenTypeIsNull()
-    {
-        static void action() => TestFrameworkV2.Assert.IsNotInstanceOfType(5, null);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void InstanceNotOfTypeShouldPassOnWrongInstance()
-    {
-        TestFrameworkV2.Assert.IsNotInstanceOfType(5L, typeof(int));
-    }
-
-    [TestMethod]
-    public void InstanceNotOfTypeShouldPassOnSubInstance()
-    {
-        TestFrameworkV2.Assert.IsNotInstanceOfType(new object(), typeof(int));
-    }
-
-    #endregion
-
-    #region Nullable Booleans tests.
-
-    [TestMethod]
-    public void IsFalseNullableBooleansShouldFailWithNull()
-    {
-        bool? nullBool = null;
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.IsFalse(nullBool));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "Assert.IsFalse failed");
-    }
-
-    [TestMethod]
-    public void IsTrueNullableBooleansShouldFailWithNull()
-    {
-        bool? nullBool = null;
-
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.IsTrue(nullBool));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "Assert.IsTrue failed");
-    }
-    #endregion
-
-    #region AreNotEqual tests.
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualType()
-    {
-        static void action() => TestFrameworkV2.Assert.AreNotEqual(null, null);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualTypeWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreNotEqual(null, null, "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualString()
-    {
-        static void action() => TestFrameworkV2.Assert.AreNotEqual("A", "A");
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualStringWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreNotEqual("A", "A", "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualStringAndCaseIgnored()
-    {
-        static void action() => TestFrameworkV2.Assert.AreNotEqual("A", "a", true);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualInt()
-    {
-        static void action() => TestFrameworkV2.Assert.AreNotEqual(1, 1);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualIntWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreNotEqual(1, 1, "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualLong()
-    {
-        static void action() => TestFrameworkV2.Assert.AreNotEqual(1L, 1L);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualLongWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreNotEqual(1L, 1L, "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualLongWithDelta()
-    {
-        static void action() => TestFrameworkV2.Assert.AreNotEqual(1L, 2L, 1L);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualDecimal()
-    {
-        static void action() => TestFrameworkV2.Assert.AreNotEqual(0.1M, 0.1M);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualDecimalWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreNotEqual(0.1M, 0.1M, "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualDecimalWithDelta()
-    {
-        static void action() => TestFrameworkV2.Assert.AreNotEqual(0.1M, 0.2M, 0.1M);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualDouble()
-    {
-        static void action() => TestFrameworkV2.Assert.AreNotEqual(0.1, 0.1);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualDoubleWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreNotEqual(0.1, 0.1, "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualDoubleWithDelta()
-    {
-        static void action() => TestFrameworkV2.Assert.AreNotEqual(0.1, 0.2, 0.1);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenFloatDouble()
-    {
-        static void action() => TestFrameworkV2.Assert.AreNotEqual(100E-2, 100E-2);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenFloatDoubleWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreNotEqual(100E-2, 100E-2, "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreNotEqualShouldFailWhenNotEqualFloatWithDelta()
-    {
-        static void action() => TestFrameworkV2.Assert.AreNotEqual(100E-2, 200E-2, 100E-2);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-    #endregion
-
-    #region AreEqual tests.
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualType()
-    {
-        static void action() => TestFrameworkV2.Assert.AreEqual(null, "string");
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualTypeWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreEqual(null, "string", "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreEqual_WithTurkishCultureAndIgnoreCase_Throws()
-    {
-        var expected = "i";
-        var actual = "I";
-        var turkishCulture = new CultureInfo("tr-TR");
-
-        // In the tr-TR culture, "i" and "I" are not considered equal when doing a case-insensitive comparison.
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreEqual(expected, actual, true, turkishCulture));
-        Assert.IsNotNull(ex);
-    }
-
-    [TestMethod]
-    public void AreEqual_WithEnglishCultureAndIgnoreCase_DoesNotThrow()
-    {
-        var expected = "i";
-        var actual = "I";
-        var englishCulture = new CultureInfo("en-EN");
-
-        // Will ignore case and won't make exeption.
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreEqual(expected, actual, true, englishCulture));
-        Assert.IsNull(ex);
-    }
-
-    [TestMethod]
-    public void AreEqual_WithEnglishCultureAndDoesNotIgnoreCase_Throws()
-    {
-        var expected = "i";
-        var actual = "I";
-        var englishCulture = new CultureInfo("en-EN");
-
-        // Won't ignore case.
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreEqual(expected, actual, false, englishCulture));
-        Assert.IsNotNull(ex);
-    }
-
-    [TestMethod]
-    public void AreEqual_WithTurkishCultureAndDoesNotIgnoreCase_Throws()
-    {
-        var expected = "i";
-        var actual = "I";
-        var turkishCulture = new CultureInfo("tr-TR");
-
-        // Won't ignore case.
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreEqual(expected, actual, false, turkishCulture));
-        Assert.IsNotNull(ex);
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualStringWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreEqual("A", "a", "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualStringAndCaseIgnored()
-    {
-        static void action() => TestFrameworkV2.Assert.AreEqual("A", "a", false);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualInt()
-    {
-        static void action() => TestFrameworkV2.Assert.AreEqual(1, 2);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualIntWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreEqual(1, 2, "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualLong()
-    {
-        static void action() => TestFrameworkV2.Assert.AreEqual(1L, 2L);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualLongWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreEqual(1L, 2L, "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualLongWithDelta()
-    {
-        static void action() => TestFrameworkV2.Assert.AreEqual(10L, 20L, 5L);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualDouble()
-    {
-        static void action() => TestFrameworkV2.Assert.AreEqual(0.1, 0.2);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualDoubleWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreEqual(0.1, 0.2, "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualDoubleWithDelta()
-    {
-        static void action() => TestFrameworkV2.Assert.AreEqual(0.1, 0.2, 0.05);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualDecimal()
-    {
-        static void action() => TestFrameworkV2.Assert.AreEqual(0.1M, 0.2M);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualDecimalWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreEqual(0.1M, 0.2M, "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualDecimalWithDelta()
-    {
-        static void action() => TestFrameworkV2.Assert.AreEqual(0.1M, 0.2M, 0.05M);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenFloatDouble()
-    {
-        static void action() => TestFrameworkV2.Assert.AreEqual(100E-2, 200E-2);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenFloatDoubleWithMessage()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.AreEqual(100E-2, 200E-2, "A Message"));
-        Assert.IsNotNull(ex);
-        StringAssert.Contains(ex.Message, "A Message");
-    }
-
-    [TestMethod]
-    public void AreEqualShouldFailWhenNotEqualFloatWithDelta()
-    {
-        static void action() => TestFrameworkV2.Assert.AreEqual(100E-2, 200E-2, 50E-2);
-        ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(TestFrameworkV2.AssertFailedException));
-    }
-
-    #endregion
-
-    #region ThrowAssertFailed tests
-    [TestMethod] // See https://github.com/dotnet/sdk/issues/25373
-    public void ThrowAssertFailedDoesNotThrowIfMessageContainsInvalidStringFormatComposite()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.ThrowAssertFailed("name", "{"));
-
-        Assert.IsNotNull(ex);
-        Assert.AreEqual(typeof(TestFrameworkV2.AssertFailedException), ex.GetType());
-        StringAssert.Contains(ex.Message, "name failed. {");
-    }
-    #endregion
-
     #region BuildUserMessage tests
     [TestMethod] // See https://github.com/dotnet/sdk/issues/25373
     public void BuildUserMessageThrowsWhenMessageContainsInvalidStringFormatComposite()
@@ -678,27 +68,6 @@ public class AssertTests
     {
         string message = TestFrameworkV2.Assert.BuildUserMessage("{");
         Assert.AreEqual("{", message);
-    }
-    #endregion
-
-    #region Inconclusive tests
-    [TestMethod] // See https://github.com/dotnet/sdk/issues/25373
-    public void InconclusiveDoesNotThrowWhenMessageContainsInvalidStringFormatCompositeAndNoArgumentsPassed()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.Inconclusive("{"));
-
-        Assert.IsNotNull(ex);
-        Assert.AreEqual(typeof(TestFrameworkV2.AssertInconclusiveException), ex.GetType());
-        StringAssert.Contains(ex.Message, "Assert.Inconclusive failed. {");
-    }
-
-    [TestMethod] // See https://github.com/dotnet/sdk/issues/25373
-    public void InconclusiveThrowsWhenMessageContainsInvalidStringFormatComposite()
-    {
-        var ex = ActionUtility.PerformActionAndReturnException(() => TestFrameworkV2.Assert.Inconclusive("{", "arg"));
-
-        Assert.IsNotNull(ex);
-        Assert.AreEqual(typeof(FormatException), ex.GetType());
     }
     #endregion
 }
